+++ src/richie/apps/search/filter_definitions/courses.py
@@ -1,5 +1,6 @@
 """Define FilterDefinition classes for the courses index."""
 from functools import reduce
+import re
 
 from django import forms
 from django.utils import timezone, translation
@@ -10,6 +11,7 @@ from cms.api import Page
 from richie.apps.core.defaults import ALL_LANGUAGES_DICT
 
 from .. import ES_CLIENT
+from ..defaults import FACET_COUNTS_DEFAULT_LIMIT, FACET_COUNTS_MAX_LIMIT
 from ..fields.array import ArrayField
 from ..indexers import ES_INDICES
 from ..utils.i18n import get_best_field_language
@@ -18,12 +20,11 @@ from .mixins import (
     ChoicesAggsMixin,
     ChoicesQueryMixin,
     NestedChoicesAggsMixin,
-    TermsAggsMixin,
     TermsQueryMixin,
 )
 
 
-class IndexableFilterDefinition(TermsAggsMixin, TermsQueryMixin, BaseFilterDefinition):
+class IndexableFilterDefinition(TermsQueryMixin, BaseFilterDefinition):
     """
     Filter definition for a terms-based filter. The choices are generated dynamically from
     the incoming facets to avoid having to hold in memory or iterate over an unbounded
@@ -63,6 +64,78 @@ class IndexableFilterDefinition(TermsAggsMixin, TermsQueryMixin, BaseFilterDefin
     def aggs_include(self):
         """Do not limit which facets are computed by default."""
         return ".*"
+
+    # pylint: disable=unused-argument
+    def get_aggs_fragment(self, queries, data, *args, **kwargs):
+        """
+        Build the aggregations as a term query that counts all the different values assigned
+        to the field.
+        """
+        # Look for an include parameter in the form data and default to the regex returned
+        # by the `aggs_include` property:
+        include = data[f"{self.name:s}_include"] or self.aggs_include
+
+        # Use all the query fragments from the queries *but* the one(s) that filter on the
+        # current filter
+        filter_fragments = [
+            clause
+            for kf_pair in queries
+            for clause in kf_pair["fragment"]
+            if kf_pair["key"] is not self.name
+        ]
+
+        # Terms aggregation will return the n top facet counts among all the values taken
+        # by this field
+        terms_aggs = {
+            self.name: {
+                # Rely on the built-in "terms" aggregation to get everything we need
+                "aggregations": {
+                    self.name: {
+                        "terms": {
+                            "field": self.term,
+                            "include": include,
+                            "min_doc_count": self.min_doc_count,
+                            # Use the default limit unless the client is requesting a specific
+                            # bunch of docs, in which case we allow for more facet counts
+                            "size": FACET_COUNTS_MAX_LIMIT
+                            if data[f"{self.name:s}_include"]
+                            else FACET_COUNTS_DEFAULT_LIMIT,
+                        }
+                    }
+                },
+                # Use all the query fragments from the queries *but* the one(s) that
+                # filter on the current filter, as it is handled by ElasticSearch for us
+                "filter": {"bool": {"must": filter_fragments}},
+            }
+        }
+
+        # Filters aggregation for values that were selected in the querystring (we must force
+        # them because they may not be in the n top facet counts but we must make sure we keep
+        # them so that they remain available as options that the user can see and unselect)
+        terms_aggs.update(
+            {
+                # Create a custom aggregation for each value selected in the querystring
+                # and matching the include regex
+                # eg `organizations@P-0001` & `organizations@P-0002`
+                "{:s}@{:s}".format(self.name, value): {
+                    "filter": {
+                        "bool": {
+                            # Use all the query fragments from the queries *but* the one(s) that
+                            # filter on the current filter: we manually add back the only one that
+                            # is relevant to the current choice.
+                            "must": [{"term": {self.term: value}}]
+                            + filter_fragments
+                        }
+                    }
+                }
+                for value in data.get(self.name, [])
+                # The Elasticsearch include regex matches exact values so we must do the same
+                # by adding ^ (resp. $) at the beginning (resp. at the end) of the pattern.
+                if re.match(f"^{include:s}$", value)
+            }
+        )
+
+        return terms_aggs
 
     def get_form_fields(self):
         """
